extract styles from CurrentWaypoint.jsx to style.js
