+++ treasureHunt/Components/CurrentWaypoint.jsx
@@ -1,18 +1,11 @@
-import { View, Text, Button, Image, StyleSheet, ScrollView, AppRegistry, Dimensions } from 'react-native';
+import { View, Text, Image, ScrollView, AppRegistry, Dimensions } from 'react-native';
 import { PROVIDER_GOOGLE } from 'react-native-maps';
 import MapView from 'react-native-maps';
 import * as Location from 'expo-location';
 import { useState, useEffect } from 'react';
 import { REACT_APP_MAPS_API_KEY } from '@env';
-
-const styles = StyleSheet.create({
-	container: {
-		flex: 1,
-		backgroundColor: '#fff',
-		alignItems: 'center',
-		justifyContent: 'center',
-	},
-});
+import { TouchableOpacity } from 'react-native-gesture-handler';
+import styles from '../css/style';
 
 const CurrentWaypoint = ({ navigation }) => {
 	const [location, setLocation] = useState(null);
@@ -56,6 +49,12 @@ const CurrentWaypoint = ({ navigation }) => {
 						// backgroundColor: 'tomato',
 					}}
 				/>
+				<TouchableOpacity style={styles.baseBtn}>
+					<Text style={styles.btnText}>Found </Text>
+				</TouchableOpacity>
+				<TouchableOpacity style={styles.baseBtn}>
+					<Text style={styles.btnText}>Find the next treasure</Text>
+				</TouchableOpacity>
 			</View>
 			<View
 				// pointerEvents='none'
@@ -76,9 +75,12 @@ const CurrentWaypoint = ({ navigation }) => {
 					provider={PROVIDER_GOOGLE}
 					apiKey={apiKey}
 					showsUserLocation={true}
-					scrollEnabled={false}
+					// scrollEnabled={false}
 					minZoomLevel={15} // default => 0
-					maxZoomLevel={16} // default => 20
+					maxZoomLevel={19} // default => 20
+					//     적당한 줌 찾기
+					//     + 줌 아웃
+					// - 줌 인
 					rotateEnabled={true}
 					mapType='satellite'
 				/>
