+++ treasureHunt/css/style.js
@@ -1,0 +1,26 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+	container: {
+		flex: 1,
+		backgroundColor: '#fff',
+		alignItems: 'center',
+		justifyContent: 'center',
+	},
+	baseBtn: {
+		paddingVertical: 20,
+		marginHorizontal: 20,
+		marginVertical: 20,
+		backgroundColor: '#7cb518',
+		alignItems: 'center',
+		justifyContent: 'center',
+		borderRadius: 50,
+	},
+
+	btnText: {
+		color: '#fff',
+		fontSize: 20,
+	},
+});
+
+export default styles;
